chore: refactored all scripts to use docopt module in src
chore: refactored all scripts to use docopt module in src
